requirements.txt
absl-py==2.1.0
ale-py==0.8.1
appdirs==1.4.4
appnope==0.1.4
asttokens==2.4.1
AutoROM==0.6.1
AutoROM.accept-rom-license==0.6.1
black==24.4.2
cachetools==5.3.3
certifi==2024.2.2
charset-normalizer==3.3.2
click==8.1.7
cloudpickle==3.0.0
comm==0.2.2
contourpy==1.2.1
cycler==0.12.1
debugpy==1.8.1
decorator==4.4.2
-e git+https://github.com/kevroi/discovery.git@b99895608e3534d02f4834a789f667b5b5715c5d#egg=discovery
docker-pycreds==0.4.0
executing==2.0.1
Farama-Notifications==0.0.4
filelock==3.13.3
fonttools==4.51.0
gitdb==4.0.11
GitPython==3.1.43
google-auth==2.29.0
google-auth-oauthlib==1.2.0
grpcio==1.62.1
gymnasium==0.29.0
idna==3.6
imageio==2.34.0
imageio-ffmpeg==0.4.9
importlib_resources==6.4.0
iniconfig==2.0.0
ipykernel==6.29.4
ipython==8.24.0
jedi==0.19.1
Jinja2==3.1.3
joblib==1.4.0
jupyter_client==8.6.1
jupyter_core==5.7.2
kiwisolver==1.4.5
Markdown==3.6
markdown-it-py==3.0.0
MarkupSafe==2.1.5
matplotlib==3.8.0
matplotlib-inline==0.1.7
mdurl==0.1.2
minigrid==2.3.1
moviepy==1.0.3
mpmath==1.3.0
mypy-extensions==1.0.0
nest-asyncio==1.6.0
networkx==3.3
numpy==1.25.2
oauthlib==3.2.2
opencv-python==4.8.1.78
packaging==24.0
pandas==2.1.0
parso==0.8.4
pathspec==0.12.1
pathtools==0.1.2
pexpect==4.9.0
pillow==10.3.0
platformdirs==4.2.1
pluggy==1.5.0
proglog==0.1.10
prompt-toolkit==3.0.43
protobuf==4.23.4
psutil==5.9.8
ptyprocess==0.7.0
pure-eval==0.2.2
pyasn1==0.6.0
pyasn1_modules==0.4.0
pygame==2.5.2
Pygments==2.17.2
pyparsing==3.1.2
pytest==8.2.0
python-dateutil==2.9.0.post0
pytz==2024.1
PyYAML==6.0.1
pyzmq==26.0.3
requests==2.31.0
requests-oauthlib==2.0.0
rich==13.7.1
rsa==4.9
scikit-learn==1.3.0
scipy==1.13.0
sentry-sdk==1.44.1
setproctitle==1.3.3
Shimmy==1.1.0
six==1.16.0
smmap==5.0.1
stable-baselines3==2.1.0
stack-data==0.6.3
sympy==1.12
tensorboard==2.15.1
tensorboard-data-server==0.7.2
threadpoolctl==3.4.0
torch==2.0.1
torchvision==0.15.2
tornado==6.4
tqdm==4.66.1
traitlets==5.14.3
typing_extensions==4.11.0
tzdata==2024.1
urllib3==2.2.1
wandb==0.15.12
wcwidth==0.2.13
Werkzeug==3.0.2
